requirements.txt
tensorflow-addons==0.20.0
numpy==1.23.5
matplotlib==3.7.1
pandas==2.0.1
scikit-learn==1.2.2
opencv-python==4.7.0.72
scipy==1.10.1
imgaug==0.4.0
Pillow==9.5.0
tqdm==4.65.0
tf2onnx==1.14.0
scikit-image==0.20.0
seaborn==0.12.2